fix(config): use u32 for millisecond fields so specta can export bindings
The tauri-specta bindings gate had never run in CI (earlier gates failed
first), hiding the fact that `generate-bindings` panics:

    Attempted to export "ClockConfig.interval" but Specta forbids exporting
    BigInt-style types (usize, isize, i64, u64, i128, u128) to avoid precision
    loss.

Two exported config fields were u64: `ClockConfig::interval` and
`NotificationsConfig::timeout`. Both are millisecond durations, and u32 already
covers ~49 days — far beyond a clock tick or a dismiss timeout. u32 is also the
existing convention in this schema (`NotificationsConfig::max_visible`,
`ThemeConfig::border_radius`/`font_size`, `OutputConfig::height` are all u32),
so the u64s were the outliers.

Preferred over setting a global BigIntExportBehavior::Number, which would let
any future u64 export as a lossy JS number silently.

Note this also means debug builds could never have started: lib.rs exports
bindings under #[cfg(debug_assertions)] with .expect(), so `tauri dev` would
panic on the same error. Release builds skip it, which is why the binary ran.

+++ src-tauri/src/config/schema.rs
@@ -155,8 +155,10 @@ pub enum KeyboardMode {
 #[serde(default)]
 pub struct ClockConfig {
     pub format: String,
-    /// Update interval in milliseconds.
-    pub interval: u64,
+    /// Update interval in milliseconds.  `u32` (not `u64`) because Specta
+    /// refuses to export BigInt-style types to TypeScript, and ~49 days of
+    /// milliseconds is far more range than a clock tick needs.
+    pub interval: u32,
     pub timezone: Option<String>,
 }
 
@@ -228,8 +230,9 @@ pub enum PowerAction {
 pub struct NotificationsConfig {
     pub position: NotificationPosition,
     pub max_visible: u32,
-    /// Dismiss timeout in milliseconds; 0 = never auto-dismiss.
-    pub timeout: u64,
+    /// Dismiss timeout in milliseconds; 0 = never auto-dismiss.  `u32` for the
+    /// same reason as `ClockConfig::interval`.
+    pub timeout: u32,
 }
 
 impl Default for NotificationsConfig {
